feat: bulk feedback for multiple trades (Issue #9) — v1.12.0 v2 - fix path
